[ITSASU-1077] Remove 'old' DateModelMapping and rename NewDateModelMapping to replace it
